Rename CGAModelArgument -> ForecasterTrainingArgument
